+++ packages/editor/src/gesture_fuzz.test.ts
@@ -64,6 +64,7 @@ function makeUI(activeTool = 'selection'): UIEngine {
         setActiveRegion() {},
         syncWithSelection() {},
         updateLayerList() {},
+        revealSelection() {},
         hideContextMenu() {},
         refreshArtboardPanel() {},
         applyToolCursor() {},