+++ packages/editor/src/history_discipline.test.ts
@@ -67,6 +67,7 @@ function makeUI(activeTool = 'selection'): UIEngine {
         setActiveRegion() {},
         syncWithSelection() {},
         updateLayerList() {},
+        revealSelection() {},
         hideContextMenu() {},
         refreshArtboardPanel() {},
         applyToolCursor() {},