+++ packages/editor/src/input_drag.test.ts
@@ -68,6 +68,7 @@ function makeUI(activeTool = 'selection'): UIEngine {
         activeTool,
         syncWithSelection() {},
         updateLayerList() {},
+        revealSelection() {},
         hideContextMenu() {},
         refreshArtboardPanel() {},
         applyToolCursor() {},