+++ packages/editor/src/input_field_chords.test.ts
@@ -62,6 +62,7 @@ function makeUI(): UIEngine {
         },
         syncWithSelection() {},
         updateLayerList() {},
+        revealSelection() {},
         hideContextMenu() {},
         refreshArtboardPanel() {},
         applyToolCursor() {},