+++ packages/editor/src/live_paint_remap.test.ts
@@ -70,6 +70,7 @@ function makeUI(activeTool = 'selection'): UIEngine {
         setActiveRegion() {},
         syncWithSelection() {},
         updateLayerList() {},
+        revealSelection() {},
         hideContextMenu() {},
         refreshArtboardPanel() {},
         applyToolCursor() {},